+++ Packages/ai.videokit.videokit/Runtime/ReplayBuffer.cs
@@ -95,10 +95,19 @@ namespace VideoKit {
             // Concatenate the previous chunk to the final one if the final one is not long enough
             var chunk = await chunkTask;
             if (result.duration < duration && chunk != null)
-                result = await MediaAsset.Concatenate(chunk, result);
+                result = await MediaAsset.Concatenate(
+                    new[] { chunk, result },
+                    format: format,
+                    prefix: prefix
+                );
             // Trim the result if longer than duration
             if (result.duration > duration)
-                result = await MediaAsset.TakeLast(result, duration);
+                result = await MediaAsset.TakeLast(
+                    result,
+                    duration: duration,
+                    format: format,
+                    prefix: prefix
+                );
             // Return
             return result;
         }